+++ src/main/java/ch/heigvd/Main.java
@@ -2,6 +2,9 @@ package ch.heigvd;
 
 import picocli.CommandLine;
 
+/**
+ * Main class for the CLI application
+ */
 public class Main {
     public static void main(String[] args) {
         int exitCode = new CommandLine(new CliApp()).execute(args);
