+++ src/main/java/ch/heigvd/Processor/IDataProcessor.java
@@ -1,6 +1,20 @@
 package ch.heigvd.Processor;
 
-public interface IDataProcessor {
+/**
+ * Interface for data processors
+ */
+public interface IDataProcessor
+{
+    /**
+     * Process the input data
+     * @param input The input data
+     * @return The processed data
+     */
     byte[] process(byte[] input);
+
+    /**
+     * Get the success message
+     * @return The success message
+     */
     String getSuccessMessage();
 }
